Move Phase enum into its own file
