+++ src/app/core/models/phase.model.ts
@@ -1,0 +1,6 @@
+export enum Phase {
+  phaseBugReporting = 'phaseBugReporting',
+  phaseTeamResponse = 'phaseTeamResponse',
+  phaseTesterResponse = 'phaseTesterResponse',
+  phaseModeration = 'phaseModeration'
+}
